Type mismatch

+++ Scene/Sphere.hpp
@@ -20,5 +20,5 @@ private:
 	uint32_t specularExponent;
 
 public:
-	Sphere(std::string name, float posX, float posY, float posZ, float scaleX, float scaleY, float scaleZ, float r, float g, float b, float ka, float kd, float ks, float kr, float specExp);
+	Sphere(std::string name, float posX, float posY, float posZ, float scaleX, float scaleY, float scaleZ, float r, float g, float b, float ka, float kd, float ks, float kr, uint32_t specExp);
 };
